backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlmodel==0.0.32
aiosqlite==0.21.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9